feat: reorganize imports in __main__.py for improved clarity and structure
